backend/requirements.txt
# UI Config 智能配置系统 - 后端依赖
# Python 3.10+

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0
greenlet>=3.0.0

# Pydantic
pydantic==2.6.1
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0
aiofiles==23.2.1

# JSON Schema Validation
jsonschema==4.21.1

# Image Processing (optional, for validation)
Pillow==10.2.0

# Testing
pytest>=7.0.0,<8.0.0
pytest-asyncio==0.23.4

# Development
python-multipart==0.0.9

